requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
base58==2.1.1
bitarray==2.9.2
blinker==1.4
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
construct==2.10.68
construct-typing==0.5.6
cytoolz==0.12.3
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.3
django-cors-headers==4.3.1
django-heroku==0.3.1
djangorestframework==3.15.1
dnspython==2.6.1
eth-account==0.11.0
eth-hash==0.7.0
eth-keyfile==0.8.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==4.0.0
eth_abi==5.0.1
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
hexbytes==0.3.1
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jsonalias==0.1.1
mongoengine==0.28.2
openai==1.16.2
packaging==24.0
parsimonious==0.9.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==4.6.2
PyNaCl==1.5.0
python-dotenv==1.0.1
regex==2023.12.25
repoze.lru==0.7
requests==2.31.0
rlp==4.0.0
scrypt==0.8.24
sniffio==1.3.1
solana==0.33.0
solders==0.21.0
sqlparse==0.4.4
toolz==0.12.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
websockets==11.0.3
whitenoise==6.6.0
